# Speedup seed scripts