Speeds up seeding using transactions and smarter file loading.

For example, on my system:
- Events 10s -> 2.6s
- Documents 9.5s -> 4.6s
- Images 4.64s -> 1.2s
- ...

Or about 70% less time.